chore: extract useMapPin hook from duplicated pin-state logic
HomeClient and ArchivesClient shared ~20 identical lines of
isMapSticky/isAnimating/togglePin state machine. Extracted to
shared/hooks/useMapPin.mjs with configurable default.
